community_projects/Project_2_News_Sentiment_Analyser/requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community
unstructured # UnstructuredURLLoader

# OpenAI Integration
langchain-openai

# Frontend UI
streamlit

# Environment Variable Management
python-dotenv

# PDF Processing
pypdf

# FAISS
faiss-cpu